fix(mcp): stop prompting for a per-vault MuninnDB API key
The MuninnDB key is instance-level, not per-vault — one registered `muninn`
MCP server reaches every vault (vault is a per-tool-call parameter). Decouple
vault-name capture from the key:

- runVaultWizard records the vault name always; only prompts for a key when no
  muninn MCP server is registered yet (shared isMuninnRegistered helper,
  extracted from the muninn-mcp doctor check). Blank key still records the name.
- writeApiKeyToEnv writes a single MUNINN_DB_API_KEY (the per-vault aliasing
  wrote one value under three names and was never read); VaultConfig.apiKey is
  optional; vault:init writes .env only when a key was captured.
- Reword prompt copy: the key is a one-time, instance-level MCP-registration
  credential, not a per-vault secret.
